Ecowave Website
Ecowave is a group web project that would allow pople to know about recycling and taking care of the environment. I helped design and program the front-end with Bootstrap, CSS, HTML, JS and SASS. I developed the back-end creating an API with Laravel working with PHP, for the database I used Heidi.SQL. This project was developed on Visual Studio Code.